requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# AI & Claude Integration
anthropic==0.7.8
openai==1.3.7
tiktoken==0.5.2

# HTTP & Requests
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
pydantic-settings==2.1.0

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
python-dotenv==1.0.0
black==23.11.0
isort==5.12.0

# Google Integration (NEW)
gspread-asyncio==1.4.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-api-python-client==2.110.0
aiofiles==23.2.1

# Additional Dependencies
python-dateutil==2.8.2
pytz==2023.3